Update yOCTProcessTiledScan.m
There was a minor issue with the way the dispersion parameter was read. 
I changed the function (lines 109-120) in a way that as long as the user did not enter a dispersion value the dispersion will be taken from the probe file.  
In case the function did not accept any dispersion value (actual value or empty array) a build in default value will be taken.